Move kgx.stream
