--- kgx/stream/__init__.py
@@ -1,1 +1,0 @@
-from .stream import Stream
